# Faster server variable handling