--- a/scripts/zones/Mhaura/Zone.lua
+++ b/scripts/zones/Mhaura/Zone.lua
@@ -16,7 +16,7 @@ zoneObject.onGameHour = function(zone)
         GetNPCByID(ID.npc.LAUGHING_BISON):setAnimationSub(0)
     end
 
-    SetServerVariable('Mhaura_Deastination', math.random(1, 100))
+    SetServerVariable('Mhaura_Destination', math.random(1, 100))
 end
 
 zoneObject.onInitialize = function(zone)
@@ -83,7 +83,7 @@ zoneObject.onEventFinish = function(player, csid, option, npc)
         local DepartureTime = VanadielHour()
 
         if DepartureTime % 8 == 0 then
-            if GetServerVariable('Mhaura_Deastination') > 89 then
+            if GetServerVariable('Mhaura_Destination') > 89 then
                 player:setPos(0, 0, 0, 0, xi.zone.SHIP_BOUND_FOR_SELBINA_PIRATES)
             else
                 player:setPos(0, 0, 0, 0, xi.zone.SHIP_BOUND_FOR_SELBINA)