@TeoTwawki 2023-10-12
Decided this is definitely NOT the direction I want to try and patch this up with. I still don't understand why the servervar queries are being this slow on having to update existing rows. There only one key here and neither the table nor the row should have a bad locking problem going on (we do 3 tries at reading it to confirm it was written, but that is going to take < 30ms in total).